RHCLOUD-24624 Remove RHOSE
