+++ .rhcicd/clowdapp-backend.yaml
@@ -192,26 +192,6 @@ objects:
           value: ${QUARKUS_OPENTELEMETRY_ENABLED}
         - name: QUARKUS_OPENTELEMETRY_TRACER_EXPORTER_OTLP_ENDPOINT
           value: ${QUARKUS_OPENTELEMETRY_ENDPOINT}
-        - name: OB_BRIDGE_NAME
-          value: ${OB_BRIDGE_NAME}
-        - name: OB_READY_CHECK_PERIOD
-          value: ${OB_READY_CHECK_PERIOD}
-        - name: OB_TOKEN_CLIENT_ID
-          valueFrom:
-            secretKeyRef:
-              name: rhose
-              key: token_client_id
-              optional: true
-        - name: OB_TOKEN_CLIENT_SECRET
-          valueFrom:
-            secretKeyRef:
-              name: rhose
-              key: token_client_secret
-              optional: true
-        - name: QUARKUS_REST_CLIENT_OB_URL
-          value: ${QUARKUS_REST_CLIENT_OB_URL}
-        - name: QUARKUS_REST_CLIENT_KC_URL
-          value: ${QUARKUS_REST_CLIENT_KC_URL}
         - name: NOTIFICATIONS_USE_SOURCES_SECRETS_BACKEND
           value: ${SOURCES_ENABLED}
         - name: NOTIFICATIONS_USE_EVENT_TYPE_FOR_SUBSCRIPTION_ENABLED
@@ -311,21 +291,9 @@ parameters:
 - name: NOTIFICATIONS_EMAILS_ONLY_MODE_ENABLED
   description: When this is true, all integration types except emails are disabled
   value: "true"
-- name: OB_BRIDGE_NAME
-  description: The name of the bridge to use.
-  value: "notifications-bridge"
-- name: OB_READY_CHECK_PERIOD
-  description: Period between two executions of the EndpointReadyChecker
-  value: "off"
 - name: QUARKUS_LOG_CLOUDWATCH_API_CALL_TIMEOUT
   description: Amount of time to allow the CloudWatch client to complete the execution of an API call expressed with the ISO-8601 duration format PnDTnHnMn.nS.
   value: PT30S
-- name: QUARKUS_REST_CLIENT_OB_URL
-  description: URL of the OpenBridge fleet manager
-  value: "http://addMe"
-- name: QUARKUS_REST_CLIENT_KC_URL
-  description: URL of the SSO server
-  value: "https://addMe"
 - name: RBAC_AUTHENTICATION_READ_TIMEOUT
   description: Delay in milliseconds before an RBAC authentication query is interrupted
   value: "2000"
