--- common/src/main/java/com/redhat/cloud/notifications/openbridge/ProcessingError.java
@@ -1,50 +1,0 @@
-package com.redhat.cloud.notifications.openbridge;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-
-import java.time.ZonedDateTime;
-import java.util.Map;
-
-import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
-
-/**
- *
- */
-@JsonIgnoreProperties(ignoreUnknown = true)
-@JsonNaming(SnakeCaseStrategy.class)
-public class ProcessingError {
-
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss.SSSSSSZ")
-    private ZonedDateTime recordedAt;
-
-    private Map<String, String> headers;
-
-    private JsonNode payload;
-
-    public ZonedDateTime getRecordedAt() {
-        return recordedAt;
-    }
-
-    public void setRecordedAt(ZonedDateTime recordedAt) {
-        this.recordedAt = recordedAt;
-    }
-
-    public Map<String, String> getHeaders() {
-        return headers;
-    }
-
-    public void setHeaders(Map<String, String> headers) {
-        this.headers = headers;
-    }
-
-    public JsonNode getPayload() {
-        return payload;
-    }
-
-    public void setPayload(JsonNode payload) {
-        this.payload = payload;
-    }
-}
